refactored code, added data utils functions
